+++ NEWS.md
@@ -1,3 +1,19 @@
+ParallelLogger 3.4.0
+====================
+
+Changes
+
+1. Exposing `getThreadNumber()` function that can be used to determine which thread we're in (including the main thread).
+
+2. Updating `selectFromList()` to support selection from R6 objects.
+
+3. Added `getPhysicalMemory()` function.
+
+4. Added `setAndromedaMemoryLimit` argument to `makeCluster()`. When TRUE, the `andromedaMemoryLimit` option will be set in each thread to be either the global `andromedaMemoryLimit / numberOfThreads` or 75% of the system memory / number of threads.
+
+5. Do not register loggers if running as notebook; doing so causes stack error.
+
+
 ParallelLogger 3.3.1
 ====================